feat(cleanup): Support fractional threshold for NaN cleanup
The threshold for dropping rows with NaN/Inf values can now be
specified as a float between 0 and 1, representing a fraction of the
total columns.

This allows for more flexible cleaning when the number of columns is
variable. The default `nan_rows` threshold has been updated from 160
to 0.5. An assertion was also added to prevent the preprocessing
from removing all data.